Add delivery arc but do not include it and correct connectivity
Forgot to consider F and E arcs in the Oroville node continuity.
I add D_OROVL to but it is not activated. The reservoir levels of
Oroville is exceeding the actual capacity because the Oroville rule
curve logic is not implemented yet. There are many connections to
delivery and demands and allocations so I will just need to stick to
this layout for now until the deliveries and demands are corrected.